Ajout de tests pour la fonction de décodage des listes et amélioration de la gestion des usines, y compris la sérialisation et la désérialisation. Mise à jour du formulaire d'ajout de produit avec des validations et des améliorations de style. Correction de la logique de stockage dans le localStorage et ajustement des méthodes de classe pour une meilleure fiabilité.

+++ app/presentation.js
@@ -1,5 +1,4 @@
 
-
 /**
  * Décode une liste d'objets exportée par le jeu pour en extraire les données associées.
  * Analyse le format tabulé avec index, nom, quantité et valeur.
@@ -13,12 +12,13 @@ function decodeEchoesListe(value) {
     let ligne = /^(\d+)\t(.+)\t(\d+)\t(.+)$/.exec(element);
     if (ligne) {
       let [, index, nom, quantité_brut, valeur_brut] = ligne;
-      let quantité = Number.parseInt(quantité_brut);
+      let quantité = Number.parseFloat(quantité_brut);
       let valeur = Number.parseFloat(valeur_brut);
       liste.push({ index, nom, quantité, valeur })
     }
   });
   return liste;
 }
+
 
 export { decodeEchoesListe }
